really fix contiguous arrays and also a warning in lib/memory.cpp snprintf %lld requires a long long int argument

+++ src/lib/memory.cpp
@@ -23,7 +23,7 @@ void* operator new(size_t size) throw (std::bad_alloc)
 		const size_t buf_len=128;
 		char buf[buf_len];
 		size_t written=snprintf(buf, buf_len,
-			"Out of memory error, tried to allocate %lld bytes using new().\n", (uintmax_t) size);
+			"Out of memory error, tried to allocate %lld bytes using new().\n", (long long int) size);
 		if (written<buf_len)
 			throw ShogunException(buf);
 		else
@@ -48,7 +48,7 @@ void* operator new[](size_t size)
 		const size_t buf_len=128;
 		char buf[buf_len];
 		size_t written=snprintf(buf, buf_len,
-			"Out of memory error, tried to allocate %lld bytes using new[].\n", (uintmax_t) size);
+			"Out of memory error, tried to allocate %lld bytes using new[].\n", (long long int) size);
 		if (written<buf_len)
 			throw ShogunException(buf);
 		else
